Recoded command execution

+++ src/com/worldcretornica/plotme/PMCommand.java
@@ -23,87 +23,133 @@ public class PMCommand implements CommandExecutor {
 	{
 		plugin = instance;
 	}
-
-	public boolean onCommand(CommandSender s, Command c, String l, String[] args) {
-		if (l.equalsIgnoreCase("plotme") || l.equalsIgnoreCase("plot") || l.equalsIgnoreCase("p")) {
-				if (args.length == 0) {
-					s.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + " v" + PlotMe.VERSION); //+ ChatColor.WHITE + " [] means optional, <> means obligated"
-					if (s instanceof Player)
-					{
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.claim", true))
-							s.sendMessage(ChatColor.RED + "/plotme claim " + ChatColor.RESET + "Claim the plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.claim.other", false))
-							s.sendMessage(ChatColor.RED + "/plotme claim <player> " + ChatColor.RESET + "Claim the plot for a player");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.auto", true))
-							s.sendMessage(ChatColor.RED + "/plotme auto " + ChatColor.RESET + "Claim the next free plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.home", true))
-							s.sendMessage(ChatColor.RED + "/plotme home[:#] " + ChatColor.RESET + "Brings you to your plot. :# if multiple plots.");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.home.other", false))
-							s.sendMessage(ChatColor.RED + "/plotme home[:#] <player> " + ChatColor.RESET + "Teleport to that player plot. :# if multiple plots.");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.info", true))
-							s.sendMessage(ChatColor.RED + "/plotme info " + ChatColor.RESET + "Displays info on the plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.comment", true))
-							s.sendMessage(ChatColor.RED + "/plotme comment <text> " + ChatColor.RESET + "Leave a comment");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.comments", true))
-							s.sendMessage(ChatColor.RED + "/plotme comments " + ChatColor.RESET + "Shows the plot comments");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.info", true))
-							s.sendMessage(ChatColor.RED + "/plotme biome " + ChatColor.RESET + "Shows current biome");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.biome", true))
-							s.sendMessage(ChatColor.RED + "/plotme biome <biome> " + ChatColor.RESET + "Sets the plot biome");
-						if (PlotMe.cPerms((Player) s, "PlotMe.use.biome", true))
-							s.sendMessage(ChatColor.RED + "/plotme biomelist " + ChatColor.RESET + "List possible biomes");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.tp", false))
-							s.sendMessage(ChatColor.RED + "/plotme tp <id> " + ChatColor.RESET + "Teleports to a plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.id", false))
-							s.sendMessage(ChatColor.RED + "/plotme id " + ChatColor.RESET + "Gets plot id and coordinates");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.clear", false))
-							s.sendMessage(ChatColor.RED + "/plotme clear " + ChatColor.RESET + "Clears the plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.reset", false))
-							s.sendMessage(ChatColor.RED + "/plotme reset " + ChatColor.RESET + "Clears the plot and removes owner");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.add", false))
-							s.sendMessage(ChatColor.RED + "/plotme add <player> " + ChatColor.RESET + "Allows a player on the plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.remove", false))
-							s.sendMessage(ChatColor.RED + "/plotme remove <player> " + ChatColor.RESET + "Removes a player on the plot");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.setowner", false))
-							s.sendMessage(ChatColor.RED + "/plotme setowner <player> " + ChatColor.RESET + "Sets the plot owner");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.move", false))
-							s.sendMessage(ChatColor.RED + "/plotme move <id-from> <id-to> " + ChatColor.RESET + "Exchanges both plots");
-						if (PlotMe.cPerms((Player) s, "PlotMe.admin.reload", false))
-							s.sendMessage(ChatColor.RED + "/plotme reload " + ChatColor.RESET + "Reloads the plugin.");
-						
-					}else{
-						s.sendMessage(ChatColor.RED + "/plotme reload " + ChatColor.RESET + "Reloads the plugin.");
-					}
-					return true;
+	
+	public boolean onCommand(CommandSender s, Command c, String l, String[] args){
+		if(l.equalsIgnoreCase("plotme") || l.equalsIgnoreCase("plot") || l.equalsIgnoreCase("p")){
+			if(!(s instanceof Player)){
+				if(args.length == 0 || args[0].equalsIgnoreCase("1")){
+					s.sendMessage(PlotMe.PREFIX + " v" + PlotMe.VERSION);
+					s.sendMessage(" ---==PlotMe Console Help Page==--- "); //+ ChatColor.WHITE + " [] means optional, <> means obligated"
+					s.sendMessage(" - /plotme reload");
+					s.sendMessage(" - Reloads the plugin and its configuration files");
 				}else{
-					
 					String a0 = args[0].toString();
-					
-					if(s instanceof Player)
-					{
+					if(!(s instanceof Player)){
+						if (a0.equalsIgnoreCase("reload")) { reload(s, args); return true;}
+					}
+				}
+			}else{
+				if(args.length == 0 || args[0].equalsIgnoreCase("1")){
+					s.sendMessage(ChatColor.RED + " ---==" + ChatColor.BLUE + "PlotMe Help Page #1" + ChatColor.RED + "==--- "); //+ ChatColor.WHITE + " [] means optional, <> means obligated"
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.claim", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme claim");
+						s.sendMessage(ChatColor.AQUA + " - Claim the current plot you are standing on");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.claim.other", false))
+						s.sendMessage(ChatColor.GREEN + " - /plotme claim <player>");
+						s.sendMessage(ChatColor.AQUA + " - Claim the current plot you are standing on for another player");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.auto", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme auto");
+						s.sendMessage(ChatColor.AQUA + " - Claim the next available free plot");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.home", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme home[:#]");
+						s.sendMessage(ChatColor.AQUA + " - Teleports you to your own plot, :# if you own multiple plots");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.home.other", false))
+						s.sendMessage(ChatColor.GREEN + " - /plotme home[:#] <player>");
+						s.sendMessage(ChatColor.AQUA + " - Teleports you to other plots, :# if other people own multiple plots");
+				}else{
+					String a0 = args[0].toString();
+					if(s instanceof Player){
 						Player p = (Player)s;
-						
 						if (a0.equalsIgnoreCase("claim")) { claim(p, args); return true;}
 						if (a0.equalsIgnoreCase("auto")) { auto(p, args); return true;	}
 						if (a0.startsWith("home") || a0.startsWith("h")) { home(p, args); return true;}
+					}
+				}
+				if(args.length == 1 && args[0].equalsIgnoreCase("2")){
+					s.sendMessage(ChatColor.RED + " ---==" + ChatColor.BLUE + "PlotMe Help Page #2" + ChatColor.RED + "==--- ");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.info", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme info");
+						s.sendMessage(ChatColor.AQUA + " - Displays information about the plot your currently standing on");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.comment", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme comment <comment>");
+						s.sendMessage(ChatColor.AQUA + " - Leave comment on the current plot");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.comments", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme comments");
+						s.sendMessage(ChatColor.AQUA + " - Lists all comments users have said about your plot");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.info", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme biomeauto");
+						s.sendMessage(ChatColor.AQUA + " - Shows the current biome in the plot");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.biome", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme biome <biome>");
+						s.sendMessage(ChatColor.AQUA + " - Changes the plots biome to the one specified");
+				}else{
+					String a0 = args[0].toString();
+					if(s instanceof Player){
+						Player p = (Player)s;
 						if (a0.equalsIgnoreCase("info") || a0.equalsIgnoreCase("i")) { info(p, args); return true;}
 						if (a0.equalsIgnoreCase("comment")) { comment(p, args); return true;}
 						if (a0.equalsIgnoreCase("comments") || a0.equalsIgnoreCase("c")) { comments(p, args); return true;}
 						if (a0.equalsIgnoreCase("biome") || a0.equalsIgnoreCase("b")) { biome(p, args); return true;}
+					}
+				}
+				if(args.length == 1 && args[0].equalsIgnoreCase("3")){
+					s.sendMessage(ChatColor.RED + " ---==" + ChatColor.BLUE + "PlotMe Help Page #3" + ChatColor.RED + "==--- ");
+					if(PlotMe.cPerms((Player) s, "PlotMe.use.biome", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme biomelist");
+						s.sendMessage(ChatColor.AQUA + " - List all possible biomes");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.tp", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme tp <id>");
+						s.sendMessage(ChatColor.AQUA + " - Teleports to a plot in the current world");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.id", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme id");
+						s.sendMessage(ChatColor.AQUA + " - Gets plot id and coordinates of the current plot your standing on");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.clear", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme clear");
+						s.sendMessage(ChatColor.AQUA + " - Clear the plot to its original flat state");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.reset", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme reset");
+						s.sendMessage(ChatColor.AQUA + " - Reset the plot to its original flat state AND remove its owner");
+				}else{
+					String a0 = args[0].toString();
+					if(s instanceof Player){
+						Player p = (Player)s;
 						if (a0.equalsIgnoreCase("biomelist")) { biomelist(p, args); return true;}
 						if (a0.equalsIgnoreCase("id")) { id(p, args); return true;}
 						if (a0.equalsIgnoreCase("tp")) { tp(p, args); return true;}
 						if (a0.equalsIgnoreCase("clear")) { clear(p, args); return true;}
 						if (a0.equalsIgnoreCase("reset")) { reset(p, args); return true;}
+					}
+				}
+				if(args.length == 1 && args[0].equalsIgnoreCase("4")){
+					s.sendMessage(ChatColor.RED + " ---==" + ChatColor.BLUE + "PlotMe Help Page #4" + ChatColor.RED + "==--- ");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.add", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme add <player>");
+						s.sendMessage(ChatColor.AQUA + " - Allows a player to have full access to the plot(This is your responsibility!)");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.remove", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme remove <player>");
+						s.sendMessage(ChatColor.AQUA + " - Revoke a players access to the plot");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.setowner", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme setowner <player>");
+						s.sendMessage(ChatColor.AQUA + " - Sets the player provided as the owner of the plot your currently on");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.move", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme move <id-from> <id-to>");
+						s.sendMessage(ChatColor.AQUA + " - Swaps the plots blocks(highly experimental for now, use at your own risk)");
+					if(PlotMe.cPerms((Player) s, "PlotMe.admin.reload", true))
+						s.sendMessage(ChatColor.GREEN + " - /plotme reload");
+						s.sendMessage(ChatColor.AQUA + " - Reloads the plugin and its configuration files");
+				}else{
+					String a0 = args[0].toString();
+					if(s instanceof Player){
+						Player p = (Player)s;
 						if (a0.equalsIgnoreCase("add") || a0.equalsIgnoreCase("+")) { add(p, args); return true;}
 						if (a0.equalsIgnoreCase("remove") || a0.equalsIgnoreCase("-")) { remove(p, args); return true;}
 						if (a0.equalsIgnoreCase("setowner") || a0.equalsIgnoreCase("o")) { setowner(p, args); return true;}
 						if (a0.equalsIgnoreCase("move") || a0.equalsIgnoreCase("m")) { move(p, args); return true;}
+						if (a0.equalsIgnoreCase("reload")) { reload(s, args); return true;}
 					}
-					if (a0.equalsIgnoreCase("reload")) { reload(s, args); return true;}
-					
 				}
-			}		
+			}
+		}
 		return false;
 	}
 	
@@ -486,25 +532,28 @@ public class PMCommand implements CommandExecutor {
 	{
 		if (PlotMe.cPerms(p, "PlotMe.admin.reset", false))
 		{
-			String id = PlotManager.getPlotId(p.getLocation());
-			if(id.equals(""))
+			if(PlotMe.checkPerms(p, "PlotMe.admin"))
 			{
-				p.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + ChatColor.RED + " No plot found");
-			}else{
-				Location bottom = PlotManager.getPlotBottomLoc(p.getWorld(), id);
-				Location top = PlotManager.getPlotTopLoc(p.getWorld(), id);
-				
-				PlotManager.clear(bottom, top);
-				
-				SqlManager.deletePlot(PlotManager.getIdX(id), PlotManager.getIdZ(id), p.getWorld().getName());
-				
-				if(!PlotManager.isPlotAvailable(id, p.getWorld().getName()))
+				String id = PlotManager.getPlotId(p.getLocation());
+				if(id.equals(""))
 				{
-					PlotMe.plotmaps.get(p.getWorld().getName()).plots.remove(id);
+					p.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + ChatColor.RED + " No plot found");
+				}else{
+					Location bottom = PlotManager.getPlotBottomLoc(p.getWorld(), id);
+					Location top = PlotManager.getPlotTopLoc(p.getWorld(), id);
+					
+					PlotManager.clear(bottom, top);
+					
+					SqlManager.deletePlot(PlotManager.getIdX(id), PlotManager.getIdZ(id), p.getWorld().getName());
+					
+					if(!PlotManager.isPlotAvailable(id, p.getWorld().getName()))
+					{
+						PlotMe.plotmaps.get(p.getWorld().getName()).plots.remove(id);
+					}
+					
+					PlotManager.removeSign(p.getWorld(), id);
+					p.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + ChatColor.WHITE + " Plot has been reset.");
 				}
-				
-				PlotManager.removeSign(p.getWorld(), id);
-				p.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + ChatColor.WHITE + " Plot has been reset.");
 			}
 		}
 	}
@@ -680,11 +729,7 @@ public class PMCommand implements CommandExecutor {
 	
 	private void reload(CommandSender s, String[] args)
 	{
-		if (!(s instanceof Player))
-		{
-			plugin.initialize();
-			s.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + ChatColor.WHITE + " reloaded successfully");
-		}else if (PlotMe.cPerms((Player) s, "PlotMe.admin.reload", false))
+		if (!(s instanceof Player) || PlotMe.cPerms((Player) s, "PlotMe.admin.reload", false))
 		{
 			plugin.initialize();
 			s.sendMessage(ChatColor.BLUE + PlotMe.PREFIX + ChatColor.WHITE + " reloaded successfully");
